moved auth into auth module
